requirements.txt
anthropic==0.25.8
google-generativeai==0.7.2
openai==1.30.0
python-dotenv==1.0.0
pypdf2==4.2.0
pydantic==2.6.4
requests==2.31.0
colorama==0.4.6